SDL2 Initial commit
